endstop and encoder example script
